Replace Cake builds with NUKE (#98)
This replaces the entire build process that previously used Cake with a new build tool called NUKE (https://nuke.build/). It solves deficiencies like the ability to easily edit the project in VS2019 and VSCode as well as getting properly functioning intellisense in both. Changes to the GitHub Actions workflow now publishes a draft release in GitHub automatically when a commit is pushed to `master` and pushing to NuGet when that release is published.